pyproject.toml
[project]
name = "the_council___research_tool"
version = "0.1.0"
description = "the_council___research_tool using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.4.1",
    "python-dotenv",
    "streamlit",
    "openai",
    "langchain",
    "langchain-community",
    "chromadb",
    "sentence-transformers",
    "pypdf",
    "python-docx",
    "openpyxl",
    "faiss-cpu",
    "unstructured[local-inference]",
    "requests"
]

[project.scripts]
the_council___research_tool = "the_council___research_tool.main:run"
run_crew = "the_council___research_tool.main:run"
train = "the_council___research_tool.main:train"
replay = "the_council___research_tool.main:replay"
test = "the_council___research_tool.main:test"
run_with_trigger = "the_council___research_tool.main:run_with_trigger"
run_summarizer_ui = "the_council___research_tool.streamlit_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
